add coffee icons to user icon
