+++ lib/widgets/components/stats/leaderboard_list_entry.dart
@@ -1,9 +1,9 @@
 import 'package:coffeecard/base/strings.dart';
 import 'package:coffeecard/base/style/colors.dart';
 import 'package:coffeecard/base/style/text_styles.dart';
-import 'package:coffeecard/widgets/components/gravatar_image.dart';
 import 'package:coffeecard/widgets/components/helpers/shimmer_builder.dart';
 import 'package:coffeecard/widgets/components/list_entry.dart';
+import 'package:coffeecard/widgets/components/user_icon.dart';
 import 'package:flutter/material.dart';
 import 'package:gap/gap.dart';
 
@@ -46,8 +46,8 @@ class LeaderboardListEntry extends StatelessWidget {
                 child: _LeaderboardRankMedal(rank),
               ),
               const Gap(16),
-              GravatarImage.small(
-                hash: name,
+              const GravatarImage.small(
+                id: 0, //FIXME: ID
               ),
               const Gap(10),
               Flexible(
