+++ pubspec.yaml
@@ -119,6 +119,7 @@ flutter:
   # To add assets to your application, add an assets section, like this:
   assets:
     - assets/
+    - assets/profile_icons/
 
   fonts:
     - family: RobotoMono
